Updated graphics pipeline templates to be more modern
